- Added swiping gesture in player to flip the art for easier left handed use (swipe left to show playlist, swipe right to flip it back) - Fixed 2 places where reused cells were being autoreleased causing crashes - Fixed iPad - size and background color of HomeAlbumViewController - Fixed iPad - UI glitch after adding first server - Fixed iPad - Folders tab not reloading at all when switching servers - Fixed (hopefully) crash from updating cache size label - Fixed (282) Tapping on left size of table does nothing - Fixed (278) Unable To View Cached Songs In Build 25 - Fixed (290) stuttering playback in background - Fixed (294) Quick albums doesn't work on iPad - Fixed (270) Video files are showing in folders - Fixed (285) Landscape mode playlist visual bug - Fixed (283) EQ landscape problems - Fixed (280) "Cache" button grayed out for no reason - Fixed (293) iPad - album list dissapears when selecting 2nd album in same folder - Fixed (276) iPad - Album Artwork Does Not Appear When Playing Begins - Fixed (238) UI glitch with folder dropdown - Fixed (300) Player playlist cog button useless when playlist shown - Fixed (279) iPad - "Now Playing" Not Active In Build 25 - Fixed (286) Playtime seconds off (formatTime: now rounds up instead of down to match Subsonic)

+++ Classes/NSString+time.h
@@ -8,7 +8,7 @@
 
 @interface NSString (time)
 
-+ (NSString *)formatTime:(float)seconds;
++ (NSString *)formatTime:(double)seconds;
 + (NSString *)relativeTime:(NSDate *)date;
 
 @end
